move git repo utils to own module
